Feat/template (#1)
* adapt from original connector
* allow empty binding value
* update Jenkins file
* use node with JDK
