+++ Jenkinsfile
@@ -1,14 +1,13 @@
 timestamps {
     ansiColor('xterm') {
-        node {
-            stage('Setup') {
+        node ('bcd-794') {
+            stage('Checkout') {
                 checkout scm
             }
-
             stage('Build') {
                 try {
                     if(env.BRANCH_NAME.equals('master')){
-                      sh "./mvnw -B clean deploy -Djvm=${env.JAVA_HOME_11}/bin/java -DaltDeploymentRepository=${env.ALT_DEPLOYMENT_REPOSITORY_SNAPSHOTS}"  
+                      sh "./mvnw -B clean install -Djvm=${env.JAVA_HOME_11}/bin/java -DaltDeploymentRepository=${env.ALT_DEPLOYMENT_REPOSITORY_SNAPSHOTS}"
                     }else{
                        sh "./mvnw -B clean verify -Djvm=${env.JAVA_HOME_11}/bin/java"
                     }
@@ -16,6 +15,9 @@ timestamps {
                 } finally {
                     junit allowEmptyResults: true, testResults: '**/target/*-reports/*.xml'
                 }
+            }
+            stage('Archive') {
+                archiveArtifacts artifacts: "target/*.zip, target/*.jar", fingerprint: true
             }
         }
     }
