+++ src/main/resources/email-templating.properties
@@ -1,6 +1,6 @@
 messaging.category=Messaging
-connectorDefinitionLabel=Email (SMTP)
-connectorDefinitionDescription=Send an email using SMTP (RFC 5321)
+connectorDefinitionLabel=Email with templating (SMTP)
+connectorDefinitionDescription=Send an email using templating and SMTP (RFC 5321)
 connexionConfigPage.pageTitle=Connection Information
 connexionConfigPage.pageDescription=Enter the connection SMTP server connection information
 host.label=SMTP Host
