+++ src/main/resources/email-templating_fr.properties
