+++ src/main/resources/email-templating_ja.properties
