+++ src/test/java/org/bonitasoft/connectors/email/templating/test/EmailConnectorValidationTest.java
@@ -11,14 +11,14 @@
  * program; if not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth
  * Floor, Boston, MA 02110-1301, USA.
  **/
-package org.bonitasoft.connectors.email.test;
+package org.bonitasoft.connectors.email.templating.test;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.bonitasoft.connectors.email.EmailConnector;
+import org.bonitasoft.connectors.email.templating.EmailConnector;
 import org.bonitasoft.engine.connector.ConnectorValidationException;
 import org.bonitasoft.engine.exception.BonitaException;
 import org.junit.Test;
